[MODIFY] create node

+++ model/dto/sensorData/response.go
@@ -11,7 +11,7 @@ type (
 		ID         uint64    `json:"id,omitempty"`
 		NodeID     uint64    `json:"node_id"`
 		NodeLabel  string    `json:"node_label"`
-		GroupTh    uint64    `json:"group_th"`
+		GroupTh    uint32    `json:"group_th"`
 		SensorCode string    `json:"sensor_code"`
 		SensorType string    `json:"sensor_type"`
 		Value      float64   `json:"value"`
